Remove disableCatdocWordWrap
Having this as an option makes no sense. We should always pass the -w
flag to catdoc to prevent it inserting line breaks every 72 characters.
With -w passed catdoc will emit a line break whenever it encounters one
in the document (e.g. at the end of a paragraph.)

Control over wether the line breaks are included in the final output is
then determined by the preserveLineBreaks option